pyproject.toml
[project]
name = "mini-vector-search-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib",
    "numpy",
    "pillow",
    "scikit-learn",
    "torch",
    "transformers",
    "faiss-gpu",  
    "tqdm",       
]
